[Release-1.3.0] PyAngel & FTRL

+++ spark-on-angel/core/src/main/scala/com/tencent/angel/spark/context/PSContext.scala
@@ -31,7 +31,7 @@ abstract class PSContext {
   private[spark] def conf: Map[String, String]
   protected def stop()
 
-  def createMatrix(rows: Int, cols: Int, t: MatrixType, rowInBlock: Int, colInBlock: Int): MatrixMeta
+  def createMatrix(rows: Int, cols: Long, t: MatrixType, rowInBlock: Int, colInBlock: Int): MatrixMeta
   def destroyMatrix(meta: MatrixMeta)
 
   def createVector(dim: Int, t: VectorType, poolCapacity: Int): PSVector
